feat(client): Follow Workflow execution chain
BREAKING CHANGE: `WorkflowClient.result()` and
`WorkflowClient.createWorkflowStub()` override for attaching to an
existing Workflow now takes an object with `workflowId` and optional
`runId` instead of 2 strings.

`workflowDefaults` added to `WorkflowClientOptions`, if provided they
will be merged with per-method and per-handle options.

`WorkflowOptions` takes a `followRuns` boolean option which defaults to
`true`, it determines whether or not the client will follow the Workflow
execution chain when getting a Workflow's result.