fix!: Improve failure details (#467)
Implements all I suggested in #462  except for more specific `source`, which I think would require passing more info to `errorToFailure`. I'd like to get to that later.

This change could be breaking: if a user is checking whether `failure.message === 'Error: orig message'`, it won't catch the new `failure.message = 'orig message'`. I think the old behavior that added the `Error: ` was unintentional—at the end of `errorToFailure`, the `instanceof Error` failed because the Error was coming from the vm, so `String(err)` was used.

Co-authored-by: Roey Berman <roey@temporal.io>