Bring gRPC retry options in line with other SDKs. (#1368)
## What was changed

* Raise initial interval to 100ms (from 20ms)
* Drop max interval to 5s (from 10s)
* Drop backoff factor to 1.7 (from 2.0)
* Raise jitter factor to 0.2 (from 0.1)

## Why?

The current retry options for the TypeScript SDK are a little wonky, as the existing initial interval is *by far* the most aggressive among all Temporal SDKs.  This brings the default parameters in line with the changes already made to the Java SDK and in flight for the Core SDK.

This also fixes a wart: the TypeScript SDK can exhaust all retries in 9.200 seconds, if the random jitters for each retry align in just the right way, and we want all SDKs to keep retrying for at least 10 seconds after the initial failure.

This advances progress on https://github.com/temporalio/features/issues/27 (aka SDK-118 in internal Jira).